Configure src as test path
Instead of putting the (unit) test files in a separate test directory I choose
to put the test file next to the file (namespace) it is testing. By doing this
we do not have to duplicate the src directory structure into the test directory
anymore. Furthermore, I think it is handy to have the tests nearby. And if you
see these tests as a kind of specification (by example) of your code, why not
put them together.